Done
Demo version is completely done. The feedback function works as it is supposed to, the IR camera and the AI will keep switching between them until they are sure a person is detected. In case the drone wants to, even when everything is done, the program can just be restarted by sending the "5" command to it. So the program is always running, but when not necessary it is not reading data from both cameras to reduce memory battery usage. As far as it looks now, human recognition is done, there might be a few optimizations after running a few tests.